docs(governance): RUNBOOK STEP 29O progress registry closeout v0
Registry-only closeout after squash-merge of PR #4711 intent compatibility
firewall v1. Binds RUNBOOK_STEP_29O_COMPLETE=true and sets NEXT_RUNBOOK_STEP
to RUNBOOK_STEP_29P without starting capital/risk/sizing or runtime effects.
