+++ docs/governance/PEAK_TRADE_AUTONOMY_RUNBOOK_PROGRESS_V1.md
@@ -16,10 +16,10 @@ SCHEDULER_RUNTIME_ALLOWED: false
 
 | Feld | Wert |
 |---|---|
-| `LAST_VERIFIED_ORIGIN_MAIN` | `0e88f334d3776b554b68fe941183008047b714c6` |
-| `LAST_VERIFIED_AT` | `2026-07-01T14:45:00Z` |
+| `LAST_VERIFIED_ORIGIN_MAIN` | `8ef254621a7203a63a1bc106ec2f12192fdb553f` |
+| `LAST_VERIFIED_AT` | `2026-07-01T02:30:00Z` |
 | `CURRENT_MAJOR_GAP_PACKAGE` | `MAJOR_GAP_COMPARISON_PROMOTION_POLICY_INPUT_BRIDGE_V0` |
-| `NEXT_RUNBOOK_STEP` | `RUNBOOK_STEP_29O` |
+| `NEXT_RUNBOOK_STEP` | `RUNBOOK_STEP_29P` |
 | `NEXT_CANONICAL_STEP` | `backtest_engine_rewire_binding` |
 | `RUNBOOK_STEP_13_IMPLEMENTED` | `true` |
 | `TRADING_SESSION_SINGLE_WRITER_IMPLEMENTED` | `true` |
@@ -120,9 +120,12 @@ SCHEDULER_RUNTIME_ALLOWED: false
 | `RUNBOOK_STEP_29N_IMPLEMENTED` | `true` |
 | `RUNBOOK_STEP_29N_COMPLETE` | `true` |
 | `RUNBOOK_STEP_29O_STARTED` | `true` |
-| `RUNBOOK_STEP_29O_IMPLEMENTATION_STARTED` | `false` |
-| `RUNBOOK_STEP_29O_COMPLETE` | `false` |
-| `STEP_29O_STARTED` | `false` |
+| `RUNBOOK_STEP_29O_IMPLEMENTATION_STARTED` | `true` |
+| `RUNBOOK_STEP_29O_IMPLEMENTED_SCOPE` | `intent_compatibility_firewall_v1` |
+| `RUNBOOK_STEP_29O_IMPLEMENTED` | `true` |
+| `RUNBOOK_STEP_29O_COMPLETE` | `true` |
+| `STEP_29O_STARTED` | `true` |
+| `INTENT_COMPATIBILITY_FIREWALL_IMPLEMENTED` | `true` |
 | `PROMOTION_ECONOMIC_GATE_BINDING_STATUS` | `PASS` |
 | `PROMOTION_CANDIDATE_STATUS` | `INELIGIBLE` |
 | `PROMOTION_CANDIDATE_ELIGIBLE` | `false` |
@@ -1154,33 +1157,35 @@ Technische Zerlegung des Runbook-Übergangs: Promotion Eligibility → vollstän
 |---|---|
 | `RUNBOOK_PHASE` | 9 |
 | `RUNBOOK_STEP_ID` | `intent_compatibility_firewall_v1` |
-| `STATUS` | `IN_PROGRESS` |
+| `STATUS` | `COMPLETE` |
 | `CANONICAL_OWNER` | `src&#47;governance&#47;intent_compatibility_firewall_v1.py` <!-- pt:ref-target-ignore --> |
-| `CANONICAL_OWNER_STATUS` | `PROPOSED_PENDING_IMPLEMENTATION` |
-| `CANONICAL_OWNER_RATIFIED` | `false` |
-| `MERGED_PRS` | `none` |
-| `MERGE_COMMITS` | `none` |
-| `DURABLE_EVIDENCE_REFS` | `/Users/frnkhrz/Documents/Peak_Trade_runtime_evidence_archive_20260520T161443Z/planning_or_validation/runbook_step29o_intent_compatibility_firewall_read_only_scope_audit_v0_20260701T141500Z (MANIFEST_VERIFY_RC=0); audit_verdict=RUNBOOK_STEP_29O_READ_ONLY_SCOPE_AUDIT_PASS; minimum_safe_slice=intent_compatibility_firewall_v1_offline_slice; STEP_29O_REGISTRY_START_ONLY=true; STEP_29O_IMPLEMENTATION_STARTED=false; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_LIVE=true; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_RUNTIME=true; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_ORDERS=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29P=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29Q=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29R=true; offline_only=true; non_authorizing=true; runtime_process_started=false; scheduler_action_performed=false; venue_access_performed=false; credential_access_performed=false; trading_action_performed=false; progress_registry_closeout_status=pending_separate_closeout; FUTURES_ONLY=true; BITCOIN_DIRECTION_ALLOWED=false; LIVE_AUTHORIZED=false; ORDERS_ALLOWED=false; SCHEDULER_RUNTIME_ALLOWED=false` |
+| `CANONICAL_OWNER_STATUS` | `IMPLEMENTED` |
+| `CANONICAL_OWNER_RATIFIED` | `true` |
+| `MERGED_PRS` | `#4711` |
+| `MERGE_COMMITS` | `8ef254621a7203a63a1bc106ec2f12192fdb553f` |
+| `DURABLE_EVIDENCE_REFS` | `/Users/frnkhrz/Documents/Peak_Trade_runtime_evidence_archive_20260520T161443Z/closeout/runbook_step29o_intent_compatibility_firewall_v1_pr_squash_merge_closeout_v0_20260701T000806Z (MANIFEST_VERIFY_RC=0); intent_compatibility_firewall_pr=4711; intent_compatibility_firewall_merge_commit=8ef254621a7203a63a1bc106ec2f12192fdb553f; intent_compatibility_firewall_squash_parent=eccc4e5e1622f64f57e637839a6b3422a9712d8c; canonical_scope=intent_compatibility_firewall_v1; INTENT_COMPATIBILITY_FIREWALL_V1_BOUND=true; STEP29O_IMPLEMENTATION_COMPLETE=true; TECHNICAL_INTENT_COMPATIBILITY_FIREWALL_STACK_COMPLETE=true; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_LIVE=true; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_RUNTIME=true; STEP_29O_COMPLETION_DOES_NOT_AUTHORIZE_ORDERS=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29P=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29Q=true; STEP_29O_COMPLETION_DOES_NOT_START_STEP_29R=true; PROMOTION_ELIGIBILITY_GRANTED=false; RUNTIME_AUTHORITY_GRANTED=false; offline_only=true; non_authorizing=true; runtime_process_started=false; scheduler_action_performed=false; venue_access_performed=false; credential_access_performed=false; trading_action_performed=false; progress_registry_closeout_status=COMPLETE; FUTURES_ONLY=true; BITCOIN_DIRECTION_ALLOWED=false; LIVE_AUTHORIZED=false; ORDERS_ALLOWED=false; SCHEDULER_RUNTIME_ALLOWED=false` |
 | `DEPENDENCIES` | RUNBOOK_STEP_29N |
-| `REMAINING_GAPS` | `intent_compatibility_firewall_v1_offline_slice_implementation_pending;progress_registry_closeout_pending` |
-| `NEXT_REQUIRED_CONTRACT` | `RUNBOOK_STEP_29O_INTENT_COMPATIBILITY_FIREWALL_V1_IMPLEMENTATION` |
+| `REMAINING_GAPS` | `capital_risk_sizing_v1_offline_slice_pending` |
+| `NEXT_REQUIRED_CONTRACT` | `RUNBOOK_STEP_29P` |
 | `AUTHORITY_LEVEL` | `NON_AUTHORITIZING` |
 | `RUNBOOK_STEP_29O_STARTED` | `true` |
-| `RUNBOOK_STEP_29O_IMPLEMENTATION_STARTED` | `false` |
-| `RUNBOOK_STEP_29O_IMPLEMENTED_SCOPE` | `none` |
-| `RUNBOOK_STEP_29O_COMPLETE` | `false` |
-| `PROGRESS_REGISTRY_CLOSEOUT_PERFORMED` | `false` |
+| `RUNBOOK_STEP_29O_IMPLEMENTATION_STARTED` | `true` |
+| `RUNBOOK_STEP_29O_IMPLEMENTED_SCOPE` | `intent_compatibility_firewall_v1` |
+| `RUNBOOK_STEP_29O_COMPLETE` | `true` |
+| `PROGRESS_REGISTRY_CLOSEOUT_PERFORMED` | `true` |
 | `RUNBOOK_STEP_29O_SCOPE` | `intent_compatibility_firewall_v1` |
 | `RUNBOOK_STEP_29O_PROCESS_CLASSIFICATION` | `ACTIVE_PROGRESS` |
-| `RUNBOOK_STEP_29O_SCOPE_CLASSIFICATION` | `PROGRESS_REGISTRY_START_ONLY` |
+| `RUNBOOK_STEP_29O_SCOPE_CLASSIFICATION` | `INTENT_COMPATIBILITY_FIREWALL_V1_OFFLINE_SLICE` |
 | `RUNBOOK_STEP_29O_CANONICAL_DEFINITION` | `INTENT_COMPATIBILITY_FIREWALL` |
 | `RUNBOOK_STEP_29O_MINIMUM_SAFE_SLICE` | `intent_compatibility_firewall_v1_offline_slice` |
-| `INTENT_COMPATIBILITY_FIREWALL_IMPLEMENTED` | `false` |
+| `INTENT_COMPATIBILITY_FIREWALL_IMPLEMENTED` | `true` |
+| `INTENT_COMPATIBILITY_FIREWALL_V1_IMPLEMENTED` | `true` |
+| `TECHNICAL_INTENT_COMPATIBILITY_FIREWALL_STACK_COMPLETE` | `true` |
 | `IMPLICIT_INTENT_CONVERSION_ALLOWED` | `false` |
 | `CANONICAL_ORDER_INTENT_DEFINED` | `false` |
 | `CAPITAL_RISK_SIZING_MATHEMATICS_CHANGED` | `false` |
 | `RUNTIME_REWIRE_PERFORMED` | `false` |
-| `NEXT_RUNBOOK_STEP` | `RUNBOOK_STEP_29O` |
+| `NEXT_RUNBOOK_STEP` | `RUNBOOK_STEP_29P` |
 | `SEPARATE_GO_REQUIRED_FOR_RUNTIME_OR_EXECUTION_EFFECT` | `true` |
 | `SEPARATE_GO_REQUIRED_FOR_OFFLINE_CANONICAL_IMPLEMENTATION` | `false` |
 | `RUNTIME_EFFECT` | `false` |
@@ -1197,8 +1202,9 @@ Technische Zerlegung des Runbook-Übergangs: Promotion Eligibility → vollstän
 | `PROMOTION_CANDIDATE_ELIGIBLE` | `false` |
 | `TECHNICAL_PROMOTION_GATE_STACK_COMPLETE` | `true` |
 | `STEP_29O_INVENTORIED_INTENT_TYPES` | `true` |
-| `STEP_29O_DEFINED_COMPATIBILITY_CONTRACTS` | `false` |
-| `STEP_29O_BLOCKS_IMPLICIT_CROSS_TYPE_CONVERSIONS` | `false` |
+| `STEP_29O_DEFINED_COMPATIBILITY_CONTRACTS` | `true` |
+| `STEP_29O_BLOCKS_IMPLICIT_CROSS_TYPE_CONVERSIONS` | `true` |
+| `STEP_29O_COMPLETION_DOES_NOT_START_STEP_29P` | `true` |
 | `STEP_29O_REFERENCES_CANONICAL_ORDER_INTENT_IDENTITY` | `true` |
 | `STEP_29O_REFERENCES_ORDER_INTENT_IDEMPOTENCY_V1` | `true` |
 | `STEP_29O_REFERENCES_CANONICAL_ORDER_LIFECYCLE_V1` | `true` |
